FIN-22 [REFACTOR]: ♻️ AnimatingVC created for inheritance
