+++ FindEmAll/Screen/QuizVC.swift
@@ -7,10 +7,8 @@
 
 import UIKit
 
-class QuizVC: UIViewController {
+class QuizVC: AnimatingVC {
     
-    private let topAnimatingView = AnimatingView(color: PokeColor.PokeRed)
-    private let bottomAnimatingView = AnimatingView(color: PokeColor.PokeGrey)
     private let inputTextfield = PokeTextfield(withSpace: true)
     private let pokeImageview = PokeImageView(frame: .zero)
     private let firstInfoview = PokeInfoView() // VC로 옮겨서 하나로 만들 수 있는지 시도해보자
@@ -34,7 +32,6 @@ class QuizVC: UIViewController {
     
     override func viewIsAppearing(_ animated: Bool) {
         super.viewIsAppearing(animated)
-        configureAnimatingViews()
         loadAnimatingView()
     }
     
@@ -104,7 +101,7 @@ class QuizVC: UIViewController {
     private func configureLayout() {
         view.backgroundColor = PokeColor.PokeBlack
         
-        view.addSubviews(topAnimatingView, bottomAnimatingView, pokeImageview)
+        view.addSubview(pokeImageview)
         infoViews = [firstInfoview, secondInfoview, thirdInfoview, fourthInfoview]
         pokeImageview.backgroundColor = PokeColor.PokeBlack
         pokeImageview.delegate = self
@@ -157,20 +154,6 @@ class QuizVC: UIViewController {
         ])
     }
     
-    private func configureAnimatingViews() {
-        NSLayoutConstraint.activate([
-            topAnimatingView.topAnchor.constraint(equalTo: view.topAnchor),
-            topAnimatingView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            topAnimatingView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            topAnimatingView.heightAnchor.constraint(equalToConstant: height),
-            
-            bottomAnimatingView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            bottomAnimatingView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            bottomAnimatingView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            bottomAnimatingView.heightAnchor.constraint(equalToConstant: height)
-        ])
-    }
-    
     //MARK: - Methods
     @objc func backButtonTapped() {
         print("데이터가 보내졌습니다.")
@@ -182,37 +165,6 @@ class QuizVC: UIViewController {
         secondInfoview.animateBack(to: originalPosition[secondInfoview]!)
         thirdInfoview.animateBack(to: originalPosition[thirdInfoview]!)
         fourthInfoview.animateBack(to: originalPosition[fourthInfoview]!)
-    }
-    
-    // create a ViewController to subclass into
-    private func loadAnimatingView() {
-        let dispatchGroup = DispatchGroup()
-        dispatchGroup.enter()
-        topAnimatingView.animate(position: .down) {
-            dispatchGroup.leave()
-        }
-        
-        dispatchGroup.enter()
-        bottomAnimatingView.animate(position: .up) {
-            dispatchGroup.leave()
-        }
-    }
-    
-    private func dismissAnimatingView() {
-        let dispatchGroup = DispatchGroup()
-        dispatchGroup.enter()
-        topAnimatingView.animate(position: .up) {
-            dispatchGroup.leave()
-        }
-        
-        dispatchGroup.enter()
-        bottomAnimatingView.animate(position: .down) {
-            dispatchGroup.leave()
-        }
-        
-        dispatchGroup.notify(queue: .main) {
-            self.navigationController?.popViewController(animated: false)
-        }
     }
     
     private func createDismissKeyboardGesture() {
